+++ src/parser/here_doc.h
@@ -1,0 +1,6 @@
+#ifndef HERE_DOC_H
+# define HERE_DOC_H
+
+void	handle_here_doc(const char *delimeter, int *in_file);
+
+#endif
